Extract delete function to separate file
